Moves the article helper to the schema directory and namespace.
