+++ src/generators/schema/author.php
@@ -8,9 +8,9 @@
 namespace Yoast\WP\SEO\Presentations\Generators\Schema;
 
 use Yoast\WP\SEO\Context\Meta_Tags_Context;
-use Yoast\WP\SEO\Helpers\Article_Helper;
 use Yoast\WP\SEO\Helpers\Image_Helper;
 use Yoast\WP\SEO\Helpers\Schema;
+use Yoast\WP\SEO\Helpers\Schema\Article_Helper;
 
 /**
  * Returns schema Person data.
