+++ tests/generators/schema/article-test.php
@@ -5,8 +5,8 @@ namespace Yoast\WP\SEO\Tests\Generators\Schema;
 use Brain\Monkey;
 use Mockery;
 use stdClass;
-use Yoast\WP\SEO\Helpers\Article_Helper;
 use Yoast\WP\SEO\Helpers\Date_Helper;
+use Yoast\WP\SEO\Helpers\Schema\Article_Helper;
 use Yoast\WP\SEO\Helpers\Schema\ID_Helper;
 use Yoast\WP\SEO\Helpers\Schema\HTML_Helper;
 use Yoast\WP\SEO\Helpers\Schema\Language_Helper;
