+++ tests/generators/schema/author-test.php
@@ -1,7 +1,7 @@
 <?php
 
-use Yoast\WP\SEO\Helpers\Article_Helper;
 use Yoast\WP\SEO\Helpers\Image_Helper;
+use Yoast\WP\SEO\Helpers\Schema\Article_Helper;
 use Yoast\WP\SEO\Helpers\Schema;
 
 use Yoast\WP\SEO\Tests\Mocks\Meta_Tags_Context;
